test for the file prime_test.py
