+++ maths/prime_test.py
@@ -1,7 +1,6 @@
 """
 prime_test(n) returns a True if n is a prime number else it returns False
 """
-import unittest
 
 def prime_test(n):
     if n <= 1:
@@ -35,29 +34,3 @@ def prime_test2(n):
     # or equal to 1, it is not prime
     else:
         return False
-
-
-class TestSuite (unittest.TestCase):
-    def test_prime_test(self):
-        """
-            checks all prime numbers between 2 up to 100.
-            Between 2 up to 100 exists 25 prime numbers!
-        """
-        counter = 0
-        for i in range(2,101):
-            if prime_test(i):
-                counter += 1
-        self.assertEqual(25,counter)
-    def test_prime_test2(self):
-        """
-            checks all prime numbers between 2 up to 100.
-            Between 2 up to 100 exists 25 prime numbers!
-        """
-        counter = 0
-        for i in range(2,101):
-            if prime_test(i):
-                counter += 1
-        self.assertEqual(25,counter)
-        
-if __name__ == "__main__":
-    unittest.main()
